Move definitions of on and onWithOptions to Tea_html
