+++ lib/js/src/tea_html.js
@@ -6,6 +6,8 @@ var Vdom = require("./vdom.js");
 var Block = require("bs-platform/lib/js/block.js");
 var Curry = require("bs-platform/lib/js/curry.js");
 var $$String = require("bs-platform/lib/js/string.js");
+var Tea_json = require("./tea_json.js");
+var Tea_result = require("./tea_result.js");
 
 function text(str) {
   return /* Text */Block.__(1, [str]);
@@ -843,6 +845,45 @@ function onMouseOut(msg) {
   return Vdom.onMsg("mouseout", msg);
 }
 
+var defaultOptions = /* record */[
+  /* stopPropagation : false */0,
+  /* preventDefault : false */0
+];
+
+function onWithOptions(eventName, options, decoder) {
+  return Vdom.onCB(eventName, "", (function ($$event) {
+                if (options[/* stopPropagation */0]) {
+                  $$event.stopPropagation();
+                }
+                if (options[/* preventDefault */1]) {
+                  $$event.preventDefault();
+                }
+                return Tea_result.result_to_option(Tea_json.Decoder[/* decodeEvent */31](decoder, $$event));
+              }));
+}
+
+function on(eventName, decoder) {
+  return onWithOptions(eventName, defaultOptions, decoder);
+}
+
+var targetValue = Tea_json.Decoder[/* at */12](/* :: */[
+      "target",
+      /* :: */[
+        "value",
+        /* [] */0
+      ]
+    ], Tea_json.Decoder[/* string */2]);
+
+var targetChecked = Tea_json.Decoder[/* at */12](/* :: */[
+      "target",
+      /* :: */[
+        "checked",
+        /* [] */0
+      ]
+    ], Tea_json.Decoder[/* bool */5]);
+
+var keyCode = Tea_json.Decoder[/* field */11]("keyCode", Tea_json.Decoder[/* int */3]);
+
 function max(value) {
   return /* Attribute */Block.__(1, [
             "",
@@ -1060,5 +1101,11 @@ exports.onMouseEnter = onMouseEnter;
 exports.onMouseLeave = onMouseLeave;
 exports.onMouseOver = onMouseOver;
 exports.onMouseOut = onMouseOut;
+exports.defaultOptions = defaultOptions;
+exports.onWithOptions = onWithOptions;
+exports.on = on;
+exports.targetValue = targetValue;
+exports.targetChecked = targetChecked;
+exports.keyCode = keyCode;
 exports.Attributes = Attributes;
-/* No side effect */
+/* targetValue Not a pure module */
