+++ lib/js/src/tea_html2.js
@@ -7,8 +7,7 @@ var Block = require("bs-platform/lib/js/block.js");
 var Bytes = require("bs-platform/lib/js/bytes.js");
 var Curry = require("bs-platform/lib/js/curry.js");
 var $$String = require("bs-platform/lib/js/string.js");
-var Tea_json = require("./tea_json.js");
-var Tea_result = require("./tea_result.js");
+var Tea_html = require("./tea_html.js");
 var Caml_string = require("bs-platform/lib/js/caml_string.js");
 
 function text(str) {
@@ -1548,45 +1547,6 @@ var onCB = Vdom.onCB;
 
 var onMsg = Vdom.onMsg;
 
-var defaultOptions = /* record */[
-  /* stopPropagation : false */0,
-  /* preventDefault : false */0
-];
-
-function onWithOptions(eventName, options, decoder) {
-  return Vdom.onCB(eventName, "", (function ($$event) {
-                if (options[/* stopPropagation */0]) {
-                  $$event.stopPropagation();
-                }
-                if (options[/* preventDefault */1]) {
-                  $$event.preventDefault();
-                }
-                return Tea_result.result_to_option(Tea_json.Decoder[/* decodeEvent */31](decoder, $$event));
-              }));
-}
-
-function on(eventName, decoder) {
-  return onWithOptions(eventName, defaultOptions, decoder);
-}
-
-var targetValue = Tea_json.Decoder[/* at */12](/* :: */[
-      "target",
-      /* :: */[
-        "value",
-        /* [] */0
-      ]
-    ], Tea_json.Decoder[/* string */2]);
-
-var targetChecked = Tea_json.Decoder[/* at */12](/* :: */[
-      "target",
-      /* :: */[
-        "checked",
-        /* [] */0
-      ]
-    ], Tea_json.Decoder[/* bool */5]);
-
-var keyCode = Tea_json.Decoder[/* field */11]("keyCode", Tea_json.Decoder[/* int */3]);
-
 function onClick(msg) {
   return Vdom.onMsg("click", msg);
 }
@@ -1702,12 +1662,12 @@ function onFocus(msg) {
 var Events = /* module */[
   /* onCB */onCB,
   /* onMsg */onMsg,
-  /* defaultOptions */defaultOptions,
-  /* onWithOptions */onWithOptions,
-  /* on */on,
-  /* targetValue */targetValue,
-  /* targetChecked */targetChecked,
-  /* keyCode */keyCode,
+  /* on */Tea_html.on,
+  /* onWithOptions */Tea_html.onWithOptions,
+  /* defaultOptions */Tea_html.defaultOptions,
+  /* targetValue */Tea_html.targetValue,
+  /* targetChecked */Tea_html.targetChecked,
+  /* keyCode */Tea_html.keyCode,
   /* onClick */onClick,
   /* onDoubleClick */onDoubleClick,
   /* onMouseDown */onMouseDown,
@@ -1840,4 +1800,4 @@ exports.title = title;
 exports.link = link;
 exports.Attributes = Attributes;
 exports.Events = Events;
-/* targetValue Not a pure module */
+/* Tea_html Not a pure module */
