+++ src/tea_html.ml
@@ -329,6 +329,32 @@ let onMouseOver msg =
 let onMouseOut msg =
   onMsg "mouseout" msg
 
+type options = {
+  stopPropagation : bool;
+  preventDefault : bool;
+}
+
+let defaultOptions = {
+  stopPropagation = false;
+  preventDefault = false;
+}
+
+let onWithOptions eventName options decoder =
+  onCB eventName "" (fun event ->
+    if options.stopPropagation then event##stopPropagation () |> ignore;
+    if options.preventDefault then event##preventDefault () |> ignore;
+    event
+    |> Tea_json.Decoder.decodeEvent decoder
+    |> Tea_result.result_to_option
+  )
+
+let on eventName decoder = onWithOptions eventName defaultOptions decoder
+
+let targetValue = Tea_json.Decoder.at ["target"; "value"] Tea_json.Decoder.string
+
+let targetChecked = Tea_json.Decoder.at ["target"; "checked"] Tea_json.Decoder.bool
+
+let keyCode = Tea_json.Decoder.field "keyCode" Tea_json.Decoder.int
 
 module Attributes = struct
